# web/api/services/product/get.go
package product

import (
	"net/http"

	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
	fiberOtel "github.com/psmarcin/fiber-opentelemetry/pkg/fiber-otel"
	"github.com/uesleicarvalhoo/go-product-showcase/pkg/trace"
	"github.com/uesleicarvalhoo/go-product-showcase/web/api/models"
)

// Get product details
// @Summary 	Get Product
// @Description Get products details
// @Tags 		Product
// @Accepts 	json
// @Produce 	json
// @Param		id			path		string				true	"the uuid of product"
// @Success		200 		{object}	domain.Product
// @Failure		400 		{object}	models.MessageJSON
// @Router			/products/{id} [get].
func (s Service) GetDetails(c *fiber.Ctx) error {
	ctx, span := trace.NewSpan(fiberOtel.FromCtx(c), "get-product-by-id")
	defer span.End()

	id, err := uuid.Parse(c.Params("id"))
	if err != nil {
		trace.AddSpanError(span, err)

		return c.Status(http.StatusUnprocessableEntity).JSON(models.NewErrorMsg(err))
	}

	trace.AddSpanTags(span, map[string]string{"product_id": id.String()})

	product, err := s.usecase.Fetch(ctx, id)
	if err != nil {
		trace.AddSpanError(span, err)

		return c.Status(http.StatusBadRequest).JSON(models.NewErrorMsg(err))
	}

	return c.Status(http.StatusOK).JSON(product)
}